Django-backend-parking-lot/requirements.txt
asgiref==3.8.1
Django==5.1.4
django-cors-headers==4.6.0
django-filter==24.3
djangorestframework==3.15.2
mysqlclient==2.1.1
sqlparse==0.5.3
typing_extensions==4.12.2
